流体方程 边界条件修改了一半

+++ twoflow_MHD/dadi.h
@@ -4,8 +4,8 @@
 #include "MHD.h"
 
 void init_solve();
-void tridag(float* a, float* b, float* c, float* r, float* utri, float* gam, int n);
-int solve(float u_in[RMAX][ZMAX], float s[RMAX][ZMAX], int itermax, float tol_test);
+void tridag(double* a, double* b, double* c, double* r, double* utri, double* gam, int n);
+int solve(double u_in[RMAX][ZMAX], double s[RMAX][ZMAX], int itermax, double tol_test);
 
 void electric_field();
 void potential_solve();
